As per Best practices put all the JS code into its respective template.js file
