+++ 24life/styles/styles.css
@@ -303,9 +303,10 @@ main .section {
 }
 
 .newsletter main .hero h1 {
-  position: relative;
+  position: absolute;
   top: 25vh;
   font-size: 20px;
+  color: var(--background-color);
 }
 
 .newsletter .section.cards-container {
@@ -328,10 +329,11 @@ main .section {
   }
 
   .newsletter main .hero h1 {
-    position: relative;
+    position: absolute;
     top: 30vh;
     margin-left: 5vw;
     font-size: 40px;
+    color: var(--background-color);
   }
 }
 
@@ -351,13 +353,12 @@ main .section {
   }
 
   .newsletter main .hero h1 {
-    position: relative;
+    position: absolute;
     top: 45vh;
     margin-left: 5vw;
     font-size: 40px;
+    color: var(--background-color);
   }
-  
-  
 }
 
 @media (min-width: 1200px) {
@@ -376,10 +377,11 @@ main .section {
   }
 
   .newsletter main .hero h1 {
-    position: relative;
+    position: absolute;
     top: 60vh;
     margin-left: 5vw;
     font-size: 40px;
+    color: var(--background-color);
   }
 
 }
